[refactor] user 및 thread 디렉토리 구조 변경
