+++ rest/src/main/java/com/waglewagle/rest/thread/ThreadService.java
@@ -2,9 +2,11 @@ package com.waglewagle.rest.thread;
 
 import com.waglewagle.rest.keyword.entity.Keyword;
 import com.waglewagle.rest.keyword.repository.KeywordRepository;
-import com.waglewagle.rest.thread.ThreadDTO.CreateThreadDTO;
-import com.waglewagle.rest.thread.ThreadDTO.CreateThreadInputDTO;
-import com.waglewagle.rest.thread.ThreadDTO.ThreadResponseDTO;
+import com.waglewagle.rest.thread.dto.ThreadDTO.CreateThreadDTO;
+import com.waglewagle.rest.thread.dto.ThreadDTO.CreateThreadInputDTO;
+import com.waglewagle.rest.thread.dto.ThreadDTO.ThreadResponseDTO;
+import com.waglewagle.rest.thread.entity.Thread;
+import com.waglewagle.rest.thread.repository.ThreadRepository;
 import com.waglewagle.rest.user.entity.User;
 import com.waglewagle.rest.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
@@ -14,7 +16,7 @@ import org.springframework.transaction.annotation.Transactional;
 import java.util.*;
 import java.util.stream.Collectors;
 
-import static com.waglewagle.rest.thread.ThreadDTO.CreateThreadDTO.from;
+import static com.waglewagle.rest.thread.dto.ThreadDTO.CreateThreadDTO.from;
 
 @Service
 @RequiredArgsConstructor
