+++ rest/src/main/java/com/waglewagle/rest/thread/dto/request/threadRequest.java
@@ -1,0 +1,4 @@
+package com.waglewagle.rest.thread.dto.request;
+
+public class threadRequest {
+}
