+++ rest/src/main/java/com/waglewagle/rest/thread/dto/response/threadResponse.java
@@ -1,0 +1,4 @@
+package com.waglewagle.rest.thread.dto.response;
+
+public class threadResponse {
+}
