+++ rest/src/main/java/com/waglewagle/rest/thread/repository/ThreadRepository.java
@@ -1,6 +1,7 @@
-package com.waglewagle.rest.thread;
+package com.waglewagle.rest.thread.repository;
 
 
+import com.waglewagle.rest.thread.entity.Thread;
 import org.springframework.data.jpa.repository.EntityGraph;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
@@ -11,13 +12,13 @@ import java.util.List;
 
 
 @Repository
-public interface ThreadRepository extends JpaRepository<Thread, Long>, CustomThreadRepository {
+public interface ThreadRepository extends JpaRepository<Thread, Long>, com.waglewagle.rest.thread.repository.custom.ThreadCustomRepository {
 
     void deleteAllByParentThreadId(Long parentThreadId);
 
     @Modifying(clearAutomatically = true)
     @Query("UPDATE Thread  t SET t.keyword.id = ?2 WHERE t.keyword.id IN ?1")
-    //TODO: 배치 사이즈 정할 수 있는 것처럼, 벌크 연산에 있어 사이즈 조절이 가능할까?
+        //TODO: 배치 사이즈 정할 수 있는 것처럼, 벌크 연산에 있어 사이즈 조절이 가능할까?
     int updateAllKeywordIdByIdInBulk(List<Long> srcIdList, Long desId);
 
     @Modifying(clearAutomatically = true)
@@ -25,11 +26,11 @@ public interface ThreadRepository extends JpaRepository<Thread, Long>, CustomThr
     int deleteAllParentThreadByKeywordIdInBulk(List<Long> keywordIdList);
 
     @Query("SELECT DISTINCT t FROM Thread t " +
-                    "LEFT JOIN FETCH t.children tc " +
-                    "LEFT JOIN FETCH t.author ta " +
-                    "LEFT JOIN FETCH tc.author tca " +
+            "LEFT JOIN FETCH t.children tc " +
+            "LEFT JOIN FETCH t.author ta " +
+            "LEFT JOIN FETCH tc.author tca " +
             "WHERE t.keyword.id = ?1 " +
-                "AND t.parentThread IS NULL")
+            "AND t.parentThread IS NULL")
     @EntityGraph(type = EntityGraph.EntityGraphType.LOAD)
     List<Thread> findThreadsByParentThreadIsNullAndKeywordId(Long keywordId);
