+++ rest/src/main/java/com/waglewagle/rest/thread/repository/custom/ThreadCustomRepository.java
@@ -1,10 +1,13 @@
-package com.waglewagle.rest.thread;
+package com.waglewagle.rest.thread.repository.custom;
+
+import com.waglewagle.rest.thread.entity.Thread;
 
 import java.util.List;
 
-public interface CustomThreadRepository {
+public interface ThreadCustomRepository {
 
 
     List<Thread> findParentThreadsInKeyword(Long keywordId);
+
     List<Thread> findChildThreads(List<Long> threadIds);
 }
