+++ rest/src/main/java/com/waglewagle/rest/user/data_object/dto/UpdateProfileResponseDTO.java
@@ -1,4 +1,4 @@
-package com.waglewagle.rest.user.dto;
+package com.waglewagle.rest.user.data_object.dto;
 
 import com.waglewagle.rest.user.entity.User;
 import lombok.Getter;
@@ -11,6 +11,7 @@ public class UpdateProfileResponseDTO {
     private String username;
     private String profileImageUrl;
     private String userId;
+
     public UpdateProfileResponseDTO(User user) {
         this.username = user.getUsername();
         this.profileImageUrl = user.getProfileImageUrl();
