+++ rest/src/main/java/com/waglewagle/rest/user/data_object/dto/UsernameLoginDTO.java
@@ -1,4 +1,4 @@
-package com.waglewagle.rest.user.dto;
+package com.waglewagle.rest.user.data_object.dto;
 
 import lombok.Getter;
 import lombok.Setter;
