+++ rest/src/main/java/com/waglewagle/rest/user/data_object/dto/request/userRequest.java
@@ -1,0 +1,4 @@
+package com.waglewagle.rest.user.data_object.dto.request;
+
+public class userRequest {
+}
