+++ rest/src/main/java/com/waglewagle/rest/user/data_object/dto/response/userResponse.java
@@ -1,0 +1,4 @@
+package com.waglewagle.rest.user.data_object.dto.response;
+
+public class userResponse {
+}
