+++ rest/src/test/java/com/waglewagle/rest/thread/ThreadServiceTest.java
@@ -3,6 +3,10 @@ package com.waglewagle.rest.thread;
 import com.waglewagle.rest.community.repository.CommunityRepository;
 import com.waglewagle.rest.community.service.CommunityService;
 import com.waglewagle.rest.keyword.service.KeywordService;
+import com.waglewagle.rest.thread.controller.ThreadController;
+import com.waglewagle.rest.thread.dto.ThreadDTO;
+import com.waglewagle.rest.thread.entity.Thread;
+import com.waglewagle.rest.thread.repository.ThreadRepository;
 import com.waglewagle.rest.user.repository.UserRepository;
 import com.waglewagle.rest.user.service.UserService;
 import org.junit.jupiter.api.Test;
